requirements.txt
certifi==2021.5.30
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector==2.2.9
numpy==1.21.0
pandas==1.2.5
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.19
Werkzeug==2.0.1
wincertstore==0.2
